+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  * Android lint honors `ignoreErrors` settings.
  * Android lint can be disabled / enabled through the DSL. When disabled, the task will simply be SKIPPED.
     `LintVital*` tasks will still run, since that is part of the build step for release builds, not the check step.
+ * Android will not warn for usage of deprecated `useJack` option anymore.
 
 # v2.1
 * **2.1.8**